+++ CHANGELOG.rst
@@ -6,8 +6,10 @@ Changelog
 x.x.x ?
 =======
 
+* Added ``tooltipSort`` parameter to PieChartv2 panel
 * Fix mappings for Table
 * Added support for AWS Cross-Account in CloudwatchMetricsTarget
+
 
 0.7.1 2024-01-12
 ================